semester 1 tamat

+++ Semester_1/Pekan_09/midPoint/main.go
@@ -5,13 +5,13 @@ import (
 )
 
 func main() {
-    var x, y, z int
+    var x, y, z float64
     var hasil bool
 
     fmt.Scan(&x, &y, &z)
 	fmt.Println("");
 
-    hasil = float64(x + y) / 2.0 == float64(z) || float64(x + z) / 2.0 == float64(y) || float64(y + z) / 2.0 == float64(x)
+    hasil = x + y / 2.0 == z || x + z / 2.0 == y || y + z / 2.0 == x
 
     fmt.Println(hasil)
 }
